backend/pyproject.toml
[tool.poetry]
name = "flux-rag"
version = "0.1.0"
description = "한국가스기술공사 생성형 AI 플랫폼 RAG 프레임워크"
authors = ["Kelly"]
readme = "README.md"
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
# Web Framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-multipart = "^0.0.18"
# LLM Providers
openai = "^1.60.0"
anthropic = "^0.43.0"
httpx = "^0.28.0"
# Embeddings & Vector
sentence-transformers = "^3.4.0"
chromadb = "^0.6.0"
# Document Processing
pymupdf = "^1.25.0"
pyhwp = "^0.6.0"
python-docx = "^1.1.0"
openpyxl = "^3.1.0"
python-pptx = "^1.0.0"
# RAG & Search
rank-bm25 = "^0.2.2"
flashrank = "^0.2.0"
langchain-text-splitters = "^0.3.0"
# Agent & MCP
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
# Database
aiosqlite = "^0.20.0"
asyncpg = "^0.30.0"
# Auth
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.2.0"
ldap3 = {version = "^2.9.0", optional = true}
# Utilities
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
numpy = "^2.2.0"
sse-starlette = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.25.0"
pytest-cov = "^6.0.0"
ruff = "^0.9.0"
mypy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
